Add files via upload
I think is 90% done.  I added a few comments in the code, you can search for clifford to find them faster.  In employee.h and productionworker.h I changed the function prototypes a bit.  You can add the variable names back if you want to.  But for the most part I think the only thing left is removing the commented out code and adding in proper pseudo code.